requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
flake8==7.3.0
mccabe==0.7.0
pycodestyle==2.14.0
pyflakes==3.4.0
sqlparse==0.5.5
psycopg2-binary